refactor: extract data root and theme tokens
