+++ packages/platform-core/src/index.ts
@@ -4,3 +4,4 @@ export * from "./contexts/CurrencyContext";
 export * from "./contexts/LayoutContext";
 export * from "./contexts/ThemeContext";
 export * from "./defaultFilterMappings";
+export { loadThemeTokens } from "./themeTokens";
