+++ packages/platform-core/src/pricing.ts
@@ -2,21 +2,9 @@ import "server-only";
 
 import type { PricingMatrix, SKU } from "@types";
 import { pricingSchema } from "@types";
-import * as fsSync from "node:fs";
 import { promises as fs } from "node:fs";
 import * as path from "node:path";
-
-function resolveDataRoot(): string {
-  let dir = process.cwd();
-  while (true) {
-    const candidate = path.join(dir, "data", "shops");
-    if (fsSync.existsSync(candidate)) return candidate;
-    const parent = path.dirname(dir);
-    if (parent === dir) break;
-    dir = parent;
-  }
-  return path.resolve(process.cwd(), "data", "shops");
-}
+import { resolveDataRoot } from "./dataRoot";
 
 let cached: PricingMatrix | null = null;
