+++ packages/platform-core/src/repositories/pages/index.server.ts
@@ -6,7 +6,7 @@ import { pageSchema, type Page } from "@types";
 import { promises as fs } from "node:fs";
 import * as path from "node:path";
 import { validateShopName } from "../../../../lib/src/validateShopName";
-import { DATA_ROOT } from "../utils";
+import { DATA_ROOT } from "../../dataRoot";
 import { nowIso } from "@shared/date";
 
 /* -------------------------------------------------------------------------- */
