+++ packages/platform-core/src/repositories/returnLogistics.server.ts
@@ -3,7 +3,7 @@ import "server-only";
 import { returnLogisticsSchema, type ReturnLogistics } from "@types";
 import { promises as fs } from "node:fs";
 import * as path from "node:path";
-import { resolveDataRoot } from "./utils";
+import { resolveDataRoot } from "../dataRoot";
 
 function logisticsPath(): string {
   return path.join(resolveDataRoot(), "..", "return-logistics.json");
